ConfigurableList now finished and changes on Graph
Now you have to "register" all the classifiers and input encoders you
want to use on the graph. This is needed to automatically set the
InstanceTemplate and DataTemplate of the Nodes, and, in the future, for
leaving the management of connections between nodes to the graph (and
this will mean, by the way: fixed point algorithms!).