Introduce the ControlledEditor & EditorController classes
When an Editor uses FXML for the view, it's convenient to pass the whole
logic handling to a controller. In order to make this easy and without
boilerplate code, you should derive your Editor from
ControlledEditor, and implement the controller as an EditorController.
By default, the fxml file must be saved in the same package of your
Editor and named YourEditorClassNameView.fxml

ConfigurableListEditor and GraphEditor have been ported to
ControlledEditor